+++ src/main/java/com/sdm/snake/body/BodyElement.java
@@ -1,4 +1,4 @@
 package com.sdm.snake.body;
 
-public record BodyElement(double posX, double posY, BodyType type) {
+public record BodyElement(int posX, int posY, BodyType type) {
 }
